+++ linera-views/src/lib.rs
@@ -64,7 +64,7 @@ pub mod journaling;
 
 /// The code for encapsulating one key_value store into another that does metric
 #[cfg(feature = "metrics")]
-pub mod metered_wrapper;
+pub mod metering;
 
 /// The code for handling big values by splitting them into several small ones.
 pub mod value_splitting;
